+++ src/Schema/AbstractSqlTools.php
@@ -72,9 +72,15 @@ abstract class AbstractSqlTools implements SqlToolsInterface
      * @param string $nativeType the SQL type
      *
      * @return array an array which contains characteristics of the type
-     *               array ( 'nativetype', 'corresponding unifiedtype', minvalue, maxvalue, minlength, maxlength, autoincrement)
+     *               array (
+     *                   0 => 'nativetype',
+     *                   1 => 'corresponding unifiedtype',
+     *                   2 => minvalue,
+     *                   3 => maxvalue,
+     *                   4 => minlength,
+     *                   5 => maxlength,
+     *                   6 => autoincrement)
      *               minvalue, maxvalue, minlength, maxlength can be null
-     *
      */
     public function getTypeInfo($nativeType)
     {