+++ tests/units/queries/queriesTestAbstract.php
@@ -148,8 +148,7 @@ abstract class queriesTestAbstract extends \Jelix\UnitTests\UnitTestCaseDb
         $resultSet->addModifier(array($this, '_callbackTest'));
 
         $list = array();
-        //foreach($resultSet as $res){
-        while ($res = $resultSet->fetch()) {
+        foreach($resultSet as $res){
             $list[] = $res;
         }
         $this->assertEquals(3, count($list), 'query return bad number of results ('.count($list).')');